Rename download to downloads (dir and component names)
