--- app/src/app/components/contents/download/index.ts
@@ -1,3 +1,0 @@
-import Download from "./Download";
-
-export {Download};
