+++ app/src/app/components/contents/downloads/index.ts
@@ -1,0 +1,3 @@
+import Downloads from "./Downloads";
+
+export { Downloads };
