+++ app/src/tests/components/contents/downloads/DownloadButton.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import DownloadButton from "../../../../app/components/contents/download/DownloadButton";
+import DownloadButton from "../../../../app/components/contents/downloads/DownloadButton";
 import appConfig from "../../../../config/appConfig";
 
 let errorOnDownload = false;
